feat: :art: Agregando modulos al proyecto de citas
Estructurando con módulos el proyecto citas
