+++ src/29-PROYECTO-CitasModulos/js/classes/App.js
@@ -1,0 +1,29 @@
+import {datosCita, nuevaCita} from '../funciones.js'; 
+import { 
+    mascotaInput, 
+    propietarioInput, 
+    telefonoInput, 
+    fechaInput,
+    horaInput,
+    sintomasInput,
+    formulario
+} from '../selectores.js';
+
+class App {
+    constructor() {
+        this.initApp();
+    }
+    initApp() {
+        mascotaInput.addEventListener('change', datosCita);
+        propietarioInput.addEventListener('change', datosCita);
+        telefonoInput.addEventListener('change', datosCita);
+        fechaInput.addEventListener('change', datosCita);
+        horaInput.addEventListener('change', datosCita);
+        sintomasInput.addEventListener('change', datosCita);
+        
+        // Formulario nuevas citas
+        formulario.addEventListener('submit', nuevaCita);
+    }
+}
+
+export default App;
