+++ src/29-PROYECTO-CitasModulos/js/classes/Citas.js
@@ -1,0 +1,18 @@
+// Classes
+class Citas {
+    constructor() {
+        this.citas = []
+    }
+    agregarCita(cita) {
+        this.citas = [...this.citas, cita];
+    }
+    editarCita(citaActualizada) {
+        this.citas = this.citas.map( cita => cita.id === citaActualizada.id ? citaActualizada : cita)
+    }
+
+    eliminarCita(id) {
+        this.citas = this.citas.filter( cita => cita.id !== id);
+    }
+}
+
+export default Citas;
