+++ src/29-PROYECTO-CitasModulos/js/selectores.js
@@ -1,0 +1,15 @@
+export const mascotaInput = document.querySelector('#mascota');
+export const propietarioInput = document.querySelector('#propietario');
+export const telefonoInput = document.querySelector('#telefono');
+export const fechaInput = document.querySelector('#fecha');
+export const horaInput = document.querySelector('#hora');
+export const sintomasInput = document.querySelector('#sintomas');
+
+// Contenedor para las citas
+export const contenedorCitas = document.querySelector('#citas');
+
+// Formulario nuevas citas
+export const formulario = document.querySelector('#nueva-cita');
+
+// Heading
+export const heading = document.querySelector('#administra');
